+++ Apromore-Custom-Plugins/Process-Discoverer-Portal-Plugin/src/main/java/org/apromore/plugin/portal/processdiscoverer/actions/FilterActionOnClearFilter.java
@@ -36,7 +36,7 @@ public class FilterActionOnClearFilter extends FilterAction {
     @Override
     public boolean execute() {
         try {
-            this.setPreviousFilterCriteria(analyst.copyCurrentFilterCriteria());
+            this.setPreActionFilterCriteria(analyst.copyCurrentFilterCriteria());
             analyst.clearFilter();
             appController.updateUI(false);
             return true;
